refactor: move ingress functions to internal/k8s
